ql-fstest: checking latest changes
development outside of this fstest site, so here only a short changelog

- add a file create time stamp
- fix compile warning (type-punned pointer)
- fix possible deadlocks when filesystem went full
- try to simplify some code in the free_space() loop
- Simplify the read thread and keep reads 20 files away from writes
- EIO (unfortunately) commonly happens while stress testing lustre. That is
not nice, but for now we have to live with that. It probably better should
become and option to ignore this error. ENOSPC will easily come up when the
filesystem fills up and when several indepedent instances of the program
(and also other test programs) compete for the remaining disk space.
So we almost silently ignore this error.